rest_api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
python-jose==3.4.0
passlib==1.7.4
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11.1
psycopg==3.2.4
psycopg2-binary==2.9.10
python-multipart==0.0.20